src/export_to_bq/requirements.txt
functions-framework==3.*
google-cloud-storage==2.7.0
google-cloud-bigquery==3.5.0
functions-framework==3.*
google-cloud-storage==2.7.0
google-cloud-bigquery==3.5.0